+++ CHANGES.rst
@@ -15,6 +15,8 @@ Version 1.2.5
  * Minor **backwards incompatibility**: ``start_payment()`` "order_key" parameter is renamed to "order".
    This only affects passing kwargs. Passing the order key works but is deprecated, pass a ``DocdataOrder`` instead.
 
+* Added ``update_docdata_order`` management command.
+* Using ``-v2`` in managements commands displays the SOAP XML conversation (via logging).
 * Fix detecting "paid" status when the customer starts multiple payment attempts, but then completes the first.
   (the root issue here is the lack of a missing global "cluster/order status" field in the API - so we have to make guesses).
 * Fix ``DocdataOrder.last_payment`` property.